{{cookiecutter.project_slug}}/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
alembic==1.4.3            # via -r requirements/base.in
appdirs==1.4.4            # via black
asgi-lifespan==1.0.1      # via -r requirements/base.in
asyncpg==0.21.0           # via -r requirements/base.in
atomicwrites==1.4.0       # via pytest
attrs==20.2.0             # via flake8-bugbear, flake8-eradicate, pytest
bcrypt==3.2.0             # via -r requirements/base.in
black==20.8b1             # via -r requirements/dev.in
certifi==2020.6.20        # via httpx, sentry-sdk
cffi==1.14.3              # via bcrypt, cryptography
click==7.1.2              # via -r requirements/base.in, black, uvicorn
colorama==0.4.4           # via -r requirements/dev.in, pytest
cryptography==3.2.1       # via -r requirements/base.in
databases==0.4.0          # via -r requirements/base.in
dnspython==2.0.0          # via email-validator
ecdsa==0.14.1             # via python-jose
email-validator==1.1.1    # via -r requirements/base.in
eradicate==2.0.0          # via flake8-eradicate
fastapi==0.62.0           # via -r requirements/base.in
flake8-bugbear==20.1.4    # via -r requirements/dev.in
flake8-eradicate==1.0.0   # via -r requirements/dev.in
flake8-mutable==1.2.0     # via -r requirements/dev.in
flake8-print==3.1.4       # via -r requirements/dev.in
flake8==3.8.4             # via -r requirements/dev.in, flake8-bugbear, flake8-eradicate, flake8-mutable, flake8-print
h11==0.11.0               # via httpcore, uvicorn
httpcore==0.12.0          # via httpx
httpx==0.16.1             # via -r requirements/dev.in
idna==2.10                # via email-validator, rfc3986
iniconfig==1.1.1          # via pytest
isort==5.6.4              # via -r requirements/dev.in
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via mako
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black, mypy
mypy==0.790               # via -r requirements/dev.in
packaging==20.4           # via pytest
pathspec==0.8.0           # via black
pluggy==0.13.1            # via pytest
prometheus-async==19.2.0  # via -r requirements/base.in
prometheus-client==0.8.0  # via -r requirements/base.in, prometheus-async, starlette-exporter
psycopg2==2.8.6           # via -r requirements/base.in
py==1.9.0                 # via pytest
pyasn1==0.4.8             # via python-jose, rsa
pycodestyle==2.6.0        # via flake8, flake8-print
pycparser==2.20           # via cffi
pydantic==1.7             # via -r requirements/base.in, fastapi
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pytest-asyncio==0.14.0    # via -r requirements/dev.in
pytest==6.1.2             # via -r requirements/dev.in, pytest-asyncio
python-dateutil==2.8.1    # via alembic
python-dotenv==0.14.0     # via -r requirements/base.in
python-editor==1.0.4      # via alembic
python-jose==3.2.0        # via -r requirements/base.in
python-multipart==0.0.5   # via -r requirements/base.in
regex==2020.10.28         # via black
rfc3986[idna2008]==1.4.0  # via httpx
rsa==4.6                  # via python-jose
sentry-sdk==0.19.4        # via -r requirements/base.in
six==1.15.0               # via bcrypt, cryptography, ecdsa, flake8-print, packaging, python-dateutil, python-jose, python-multipart, sqlalchemy-utils, structlog
sniffio==1.2.0            # via asgi-lifespan, httpcore, httpx
sqlalchemy-utils==0.36.8  # via -r requirements/base.in
sqlalchemy==1.3.20        # via -r requirements/base.in, alembic, databases, sqlalchemy-utils
starlette-exporter==0.5.0  # via -r requirements/base.in
starlette==0.13.6         # via fastapi, starlette-exporter
structlog==20.1.0         # via -r requirements/base.in
toml==0.10.2              # via black, pytest
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.3  # via black, mypy
urllib3==1.25.11          # via sentry-sdk
uvicorn==0.12.2           # via -r requirements/base.in
wrapt==1.12.1             # via prometheus-async
